mm: Use pagevec ->nr to flag dedup page indices
Truncation relies on the ->index member of struct page to know what
entry to remove from the radix tree. This is a problem for dedup pages,
since they are shared between indices.

To solve this, I note that pagevec's are never larger than PAGEVEC_SIZE,
which is 14. Since the ->nr field is unsigned long, that leaves us more
than enough bits to have a flag for each page in the pagevec.

When find_get_pages() encounters a dedup page, it stores the index of
the page in the ->pages instead of a pointer to it, and indicates this
by setting the appropriate flag bit in ->nr.