mm: Add support for mmap() of dedup pages
In the MAP_PRIVATE path, simply skip a check against ->mapping. There is
an ext4-specific change here: ext4 uses a pagevec to check for dirtied
pages in the read fault path of a MAP_PRIVATE mapping, and needs to
ignore dedup pages.

In the MAP_SHARED path, simply COW the page, and remap in the new page.